% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medias_Jogadores.R
\name{medias_Jogadores}
\alias{medias_Jogadores}
\title{Médias dos jogadores
Função que calcula as médias individuais para os dez jogadores de uma determinada partida.
Para utilizar: medias_Jogadores()}
\usage{
medias_Jogadores(url_jogador)
}
\description{
Médias dos jogadores
Função que calcula as médias individuais para os dez jogadores de uma determinada partida.
Para utilizar: medias_Jogadores()
}
