% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medias_Times.R
\name{medias_Times}
\alias{medias_Times}
\title{Médias dos times.
Função que calcula as médias de cada estatística para os dois times de uma determinada partida.
Para utilizar: medias_Times()}
\usage{
medias_Times(url_partida)
}
\description{
Médias dos times.
Função que calcula as médias de cada estatística para os dois times de uma determinada partida.
Para utilizar: medias_Times()
}
